moved panel functions into their own file
